requirements.txt
# ── Core ─────────────────────────────────────────────────────────────────────
anthropic>=0.40.0
python-dotenv>=1.0.0

# ── Notebook ──────────────────────────────────────────────────────────────────
jupyter>=1.0.0
ipykernel>=6.0.0

# ── RAG (Lezione 4) ───────────────────────────────────────────────────────────
chromadb>=0.5.0
sentence-transformers>=3.0.0
pypdf>=4.0.0

# ── UI Web (Lezione 6) ────────────────────────────────────────────────────────
streamlit>=1.35.0

# ── Valutazione (Lezione 6) ───────────────────────────────────────────────────
ragas>=0.1.0

# ── Utility ───────────────────────────────────────────────────────────────────
requests>=2.31.0
